chore: Refactored is name allowed (#28954)
