+++ app/server/appsmith-server/src/main/java/com/appsmith/server/layouts/UpdateLayoutService.java
@@ -1,0 +1,3 @@
+package com.appsmith.server.layouts;
+
+public interface UpdateLayoutService extends UpdateLayoutServiceCE {}
