+++ app/server/appsmith-server/src/main/java/com/appsmith/server/newactions/refactors/NewActionRefactoringServiceCEImpl.java
@@ -3,9 +3,11 @@ package com.appsmith.server.newactions.refactors;
 import com.appsmith.external.constants.AnalyticsEvents;
 import com.appsmith.external.models.ActionConfiguration;
 import com.appsmith.external.models.ActionDTO;
+import com.appsmith.external.models.CreatorContextType;
 import com.appsmith.external.models.MustacheBindingToken;
 import com.appsmith.external.models.PluginType;
 import com.appsmith.server.configurations.InstanceConfig;
+import com.appsmith.server.constants.FieldName;
 import com.appsmith.server.domains.NewAction;
 import com.appsmith.server.dtos.EntityType;
 import com.appsmith.server.dtos.RefactorEntityNameDTO;
@@ -20,6 +22,8 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
 import org.springframework.util.StringUtils;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
@@ -125,6 +129,37 @@ public class NewActionRefactoringServiceCEImpl implements EntityRefactoringServi
                     return newActionService.updateUnpublishedAction(action.getId(), action);
                 })
                 .then();
+    }
+
+    @Override
+    public Flux<String> getExistingEntityNames(String contextId, CreatorContextType contextType, String layoutId) {
+        MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
+        if (contextId != null) {
+            params.add(FieldName.PAGE_ID, contextId);
+        }
+
+        return newActionService
+                .getUnpublishedActions(params)
+                .flatMap(actionDTO -> {
+                    /*
+                       This is unexpected. Every action inside a JS collection should have a collectionId.
+                       But there are a few documents found for plugin type JS inside newAction collection that don't have any collectionId.
+                       The reason could be due to the lack of transactional behaviour when multiple inserts/updates that take place
+                       during JS action creation. A detailed RCA is documented here
+                       https://www.notion.so/appsmith/RCA-JSObject-name-already-exists-Please-use-a-different-name-e09c407f0ddb4653bd3974f3703408e6
+                    */
+                    if (actionDTO.getPluginType().equals(PluginType.JS)
+                            && !StringUtils.hasLength(actionDTO.getCollectionId())) {
+                        log.debug(
+                                "JS Action with Id: {} doesn't have any collection Id under pageId: {}",
+                                actionDTO.getId(),
+                                contextId);
+                        return Mono.empty();
+                    } else {
+                        return Mono.just(actionDTO);
+                    }
+                })
+                .map(ActionDTO::getValidName);
     }
 
     private Mono<Set<String>> refactorNameInAction(
