+++ app/server/appsmith-server/src/main/java/com/appsmith/server/refactors/applications/RefactoringSolutionCE.java
@@ -7,4 +7,6 @@ import reactor.core.publisher.Mono;
 public interface RefactoringSolutionCE {
 
     Mono<LayoutDTO> refactorEntityName(RefactorEntityNameDTO refactorEntityNameDTO, String branchName);
+
+    Mono<Boolean> isNameAllowed(String pageId, String layoutId, String newName);
 }
