+++ app/server/appsmith-server/src/main/java/com/appsmith/server/refactors/entities/EntityRefactoringServiceCE.java
@@ -1,9 +1,11 @@
 package com.appsmith.server.refactors.entities;
 
 import com.appsmith.external.constants.AnalyticsEvents;
+import com.appsmith.external.models.CreatorContextType;
 import com.appsmith.server.dtos.EntityType;
 import com.appsmith.server.dtos.RefactorEntityNameDTO;
 import com.appsmith.server.dtos.RefactoringMetaDTO;
+import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 public interface EntityRefactoringServiceCE<T> {
@@ -21,4 +23,8 @@ public interface EntityRefactoringServiceCE<T> {
             RefactorEntityNameDTO refactorEntityNameDTO, RefactoringMetaDTO refactoringMetaDTO);
 
     Mono<Void> updateRefactoredEntity(RefactorEntityNameDTO refactorEntityNameDTO, String branchName);
+
+    default Flux<String> getExistingEntityNames(String contextId, CreatorContextType contextType, String layoutId) {
+        return Flux.empty();
+    }
 }
