+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/LayoutActionServiceImpl.java
@@ -1,15 +1,14 @@
 package com.appsmith.server.services;
 
-import com.appsmith.server.actioncollections.base.ActionCollectionService;
 import com.appsmith.server.datasources.base.DatasourceService;
 import com.appsmith.server.helpers.ResponseUtils;
+import com.appsmith.server.layouts.UpdateLayoutService;
 import com.appsmith.server.newactions.base.NewActionService;
 import com.appsmith.server.newpages.base.NewPageService;
-import com.appsmith.server.onload.internal.OnLoadExecutablesUtil;
+import com.appsmith.server.refactors.applications.RefactoringSolution;
 import com.appsmith.server.services.ce.LayoutActionServiceCEImpl;
 import com.appsmith.server.solutions.ActionPermission;
 import com.appsmith.server.solutions.PagePermission;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
@@ -18,30 +17,23 @@ import org.springframework.stereotype.Service;
 public class LayoutActionServiceImpl extends LayoutActionServiceCEImpl implements LayoutActionService {
 
     public LayoutActionServiceImpl(
-            ObjectMapper objectMapper,
             AnalyticsService analyticsService,
             NewPageService newPageService,
             NewActionService newActionService,
-            OnLoadExecutablesUtil pageLoadActionsUtil,
-            SessionUserService sessionUserService,
-            ActionCollectionService actionCollectionService,
+            RefactoringSolution refactoringSolution,
             CollectionService collectionService,
-            ApplicationService applicationService,
+            UpdateLayoutService updateLayoutService,
             ResponseUtils responseUtils,
             DatasourceService datasourceService,
             PagePermission pagePermission,
             ActionPermission actionPermission) {
-
         super(
-                objectMapper,
                 analyticsService,
                 newPageService,
                 newActionService,
-                pageLoadActionsUtil,
-                sessionUserService,
-                actionCollectionService,
+                refactoringSolution,
                 collectionService,
-                applicationService,
+                updateLayoutService,
                 responseUtils,
                 datasourceService,
                 pagePermission,
