+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/LayoutCollectionServiceImpl.java
@@ -2,8 +2,10 @@ package com.appsmith.server.services;
 
 import com.appsmith.server.actioncollections.base.ActionCollectionService;
 import com.appsmith.server.helpers.ResponseUtils;
+import com.appsmith.server.layouts.UpdateLayoutService;
 import com.appsmith.server.newactions.base.NewActionService;
 import com.appsmith.server.newpages.base.NewPageService;
+import com.appsmith.server.refactors.applications.RefactoringSolution;
 import com.appsmith.server.repositories.ActionCollectionRepository;
 import com.appsmith.server.services.ce.LayoutCollectionServiceCEImpl;
 import com.appsmith.server.solutions.ActionPermission;
@@ -18,6 +20,8 @@ public class LayoutCollectionServiceImpl extends LayoutCollectionServiceCEImpl i
     public LayoutCollectionServiceImpl(
             NewPageService newPageService,
             LayoutActionService layoutActionService,
+            UpdateLayoutService updateLayoutService,
+            RefactoringSolution refactoringSolution,
             ActionCollectionService actionCollectionService,
             NewActionService newActionService,
             AnalyticsService analyticsService,
@@ -28,6 +32,8 @@ public class LayoutCollectionServiceImpl extends LayoutCollectionServiceCEImpl i
         super(
                 newPageService,
                 layoutActionService,
+                updateLayoutService,
+                refactoringSolution,
                 actionCollectionService,
                 newActionService,
                 analyticsService,
