+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ce/LayoutCollectionServiceCEImpl.java
@@ -15,8 +15,10 @@ import com.appsmith.server.exceptions.AppsmithError;
 import com.appsmith.server.exceptions.AppsmithException;
 import com.appsmith.server.helpers.DefaultResourcesUtils;
 import com.appsmith.server.helpers.ResponseUtils;
+import com.appsmith.server.layouts.UpdateLayoutService;
 import com.appsmith.server.newactions.base.NewActionService;
 import com.appsmith.server.newpages.base.NewPageService;
+import com.appsmith.server.refactors.applications.RefactoringSolution;
 import com.appsmith.server.repositories.ActionCollectionRepository;
 import com.appsmith.server.services.AnalyticsService;
 import com.appsmith.server.services.LayoutActionService;
@@ -48,6 +50,8 @@ public class LayoutCollectionServiceCEImpl implements LayoutCollectionServiceCE 
 
     private final NewPageService newPageService;
     private final LayoutActionService layoutActionService;
+    private final UpdateLayoutService updateLayoutService;
+    private final RefactoringSolution refactoringSolution;
     private final ActionCollectionService actionCollectionService;
     private final NewActionService newActionService;
     private final AnalyticsService analyticsService;
@@ -89,7 +93,7 @@ public class LayoutCollectionServiceCEImpl implements LayoutCollectionServiceCE 
         return pageMono.flatMap(page -> {
                     Layout layout = page.getUnpublishedPage().getLayouts().get(0);
                     // Check against widget names and action names
-                    return layoutActionService.isNameAllowed(page.getId(), layout.getId(), collection.getName());
+                    return refactoringSolution.isNameAllowed(page.getId(), layout.getId(), collection.getName());
                 })
                 .flatMap(isNameAllowed -> {
                     // If the name is allowed, return list of actionDTOs for further processing
@@ -330,8 +334,8 @@ public class LayoutCollectionServiceCEImpl implements LayoutCollectionServiceCE 
                                 // 2. Run updateLayout on the old page
                                 return Flux.fromIterable(page.getLayouts())
                                         .flatMap(layout -> {
-                                            layout.setDsl(layoutActionService.unescapeMongoSpecialCharacters(layout));
-                                            return layoutActionService.updateLayout(
+                                            layout.setDsl(updateLayoutService.unescapeMongoSpecialCharacters(layout));
+                                            return updateLayoutService.updateLayout(
                                                     page.getId(), page.getApplicationId(), layout.getId(), layout);
                                         })
                                         .collect(toSet());
@@ -349,8 +353,8 @@ public class LayoutCollectionServiceCEImpl implements LayoutCollectionServiceCE 
                                 // 3. Run updateLayout on the new page.
                                 return Flux.fromIterable(page.getLayouts())
                                         .flatMap(layout -> {
-                                            layout.setDsl(layoutActionService.unescapeMongoSpecialCharacters(layout));
-                                            return layoutActionService.updateLayout(
+                                            layout.setDsl(updateLayoutService.unescapeMongoSpecialCharacters(layout));
+                                            return updateLayoutService.updateLayout(
                                                     page.getId(), page.getApplicationId(), layout.getId(), layout);
                                         })
                                         .collect(toSet());
