+++ app/server/appsmith-server/src/main/java/com/appsmith/server/solutions/CreateDBTablePageSolutionImpl.java
@@ -4,6 +4,7 @@ import com.appsmith.server.datasources.base.DatasourceService;
 import com.appsmith.server.datasourcestorages.base.DatasourceStorageService;
 import com.appsmith.server.helpers.PluginExecutorHelper;
 import com.appsmith.server.helpers.ResponseUtils;
+import com.appsmith.server.layouts.UpdateLayoutService;
 import com.appsmith.server.newpages.base.NewPageService;
 import com.appsmith.server.plugins.base.PluginService;
 import com.appsmith.server.services.AnalyticsService;
@@ -19,12 +20,12 @@ import org.springframework.stereotype.Service;
 @Slf4j
 public class CreateDBTablePageSolutionImpl extends CreateDBTablePageSolutionCEImpl
         implements CreateDBTablePageSolution {
-
     public CreateDBTablePageSolutionImpl(
             DatasourceService datasourceService,
             DatasourceStorageService datasourceStorageService,
             NewPageService newPageService,
             LayoutActionService layoutActionService,
+            UpdateLayoutService updateLayoutService,
             ApplicationPageService applicationPageService,
             ApplicationService applicationService,
             PluginService pluginService,
@@ -42,6 +43,7 @@ public class CreateDBTablePageSolutionImpl extends CreateDBTablePageSolutionCEIm
                 datasourceStorageService,
                 newPageService,
                 layoutActionService,
+                updateLayoutService,
                 applicationPageService,
                 applicationService,
                 pluginService,
