+++ app/server/appsmith-server/src/test/java/com/appsmith/server/solutions/ce/RefactoringSolutionCEImplTest.java
@@ -16,6 +16,7 @@ import com.appsmith.server.dtos.PageDTO;
 import com.appsmith.server.dtos.RefactorEntityNameDTO;
 import com.appsmith.server.exceptions.AppsmithError;
 import com.appsmith.server.helpers.ResponseUtils;
+import com.appsmith.server.layouts.UpdateLayoutService;
 import com.appsmith.server.newactions.base.NewActionService;
 import com.appsmith.server.newpages.base.NewPageService;
 import com.appsmith.server.refactors.applications.RefactoringSolutionCEImpl;
@@ -23,7 +24,6 @@ import com.appsmith.server.refactors.entities.EntityRefactoringService;
 import com.appsmith.server.repositories.ActionCollectionRepository;
 import com.appsmith.server.services.AnalyticsService;
 import com.appsmith.server.services.ApplicationService;
-import com.appsmith.server.services.LayoutActionService;
 import com.appsmith.server.services.SessionUserService;
 import com.appsmith.server.solutions.ActionPermission;
 import com.appsmith.server.solutions.PagePermission;
@@ -73,7 +73,7 @@ class RefactoringSolutionCEImplTest {
     private ResponseUtils responseUtils;
 
     @MockBean
-    private LayoutActionService layoutActionService;
+    private UpdateLayoutService updateLayoutService;
 
     @MockBean
     private ApplicationService applicationService;
@@ -110,7 +110,7 @@ class RefactoringSolutionCEImplTest {
         refactoringSolutionCE = new RefactoringSolutionCEImpl(
                 newPageService,
                 responseUtils,
-                layoutActionService,
+                updateLayoutService,
                 applicationService,
                 pagePermission,
                 analyticsService,
@@ -140,9 +140,6 @@ class RefactoringSolutionCEImplTest {
         oldActionCollection.setUnpublishedCollection(oldUnpublishedCollection);
         oldUnpublishedCollection.setDefaultResources(setDefaultResources(oldUnpublishedCollection));
         oldActionCollection.setDefaultResources(setDefaultResources(oldActionCollection));
-
-        Mockito.when(layoutActionService.isNameAllowed(Mockito.any(), Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.just(true));
 
         LayoutDTO layout = new LayoutDTO();
         final JSONObject jsonObject = new JSONObject();
@@ -182,9 +179,13 @@ class RefactoringSolutionCEImplTest {
         Mockito.when(newActionService.findByPageIdAndViewMode(Mockito.anyString(), Mockito.anyBoolean(), Mockito.any()))
                 .thenReturn(Flux.empty());
 
-        Mockito.when(layoutActionService.updateLayout(
+        Mockito.when(updateLayoutService.updateLayout(
                         Mockito.anyString(), Mockito.anyString(), Mockito.anyString(), Mockito.any()))
                 .thenReturn(Mono.just(layout));
+
+        Mockito.doReturn(Flux.just(oldUnpublishedCollection.getName()))
+                .when(actionCollectionEntityRefactoringService)
+                .getExistingEntityNames(Mockito.anyString(), Mockito.any(), Mockito.anyString());
 
         final Mono<LayoutDTO> layoutDTOMono =
                 refactoringSolutionCE.refactorEntityName(refactorActionCollectionNameDTO, null);
@@ -219,20 +220,24 @@ class RefactoringSolutionCEImplTest {
         duplicateUnpublishedCollection.setName("newName");
         duplicateActionCollection.setUnpublishedCollection(duplicateUnpublishedCollection);
 
-        Mockito.when(actionCollectionRepository.findAllActionCollectionsByNamePageIdsViewModeAndBranch(
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.anyBoolean(),
-                        Mockito.any(),
-                        Mockito.any(),
-                        Mockito.any()))
-                .thenReturn(Flux.just(oldActionCollection, duplicateActionCollection));
-
-        Mockito.when(layoutActionService.isNameAllowed(Mockito.any(), Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.just(false));
+        Mockito.doReturn(Flux.just("oldName", "newName"))
+                .when(actionCollectionEntityRefactoringService)
+                .getExistingEntityNames(Mockito.anyString(), Mockito.any(), Mockito.anyString());
 
         NewPage newPage = new NewPage();
         newPage.setId("testPageId");
+        PageDTO pageDTO = new PageDTO();
+        pageDTO.setId("testPageId");
+        pageDTO.setApplicationId("testAppId");
+        Layout layout1 = new Layout();
+        layout1.setId("testLayoutId");
+        final JSONObject jsonObject = new JSONObject();
+        jsonObject.put("key", "value");
+        layout1.setDsl(jsonObject);
+        pageDTO.setLayouts(List.of(layout1));
+        newPage.setUnpublishedPage(pageDTO);
+        Mockito.when(newPageService.findPageById(Mockito.anyString(), Mockito.any(), Mockito.anyBoolean()))
+                .thenReturn(Mono.just(pageDTO));
         Mockito.when(newPageService.getById(Mockito.anyString())).thenReturn(Mono.just(newPage));
 
         final Mono<LayoutDTO> layoutDTOMono =
@@ -264,9 +269,6 @@ class RefactoringSolutionCEImplTest {
         oldActionCollection.setUnpublishedCollection(oldUnpublishedCollection);
         oldActionCollection.setDefaultResources(setDefaultResources(oldActionCollection));
         oldUnpublishedCollection.setDefaultResources(setDefaultResources(oldUnpublishedCollection));
-
-        Mockito.when(layoutActionService.isNameAllowed(Mockito.any(), Mockito.any(), Mockito.any()))
-                .thenReturn(Mono.just(true));
 
         Mockito.when(newActionService.findActionDTObyIdAndViewMode(Mockito.any(), Mockito.anyBoolean(), Mockito.any()))
                 .thenReturn(Mono.just(new ActionDTO()));
@@ -319,9 +321,13 @@ class RefactoringSolutionCEImplTest {
         Mockito.when(responseUtils.updateLayoutDTOWithDefaultResources(Mockito.any()))
                 .thenReturn(layout);
 
-        Mockito.when(layoutActionService.updateLayout(
+        Mockito.when(updateLayoutService.updateLayout(
                         Mockito.anyString(), Mockito.anyString(), Mockito.anyString(), Mockito.any()))
                 .thenReturn(Mono.just(layout));
+
+        Mockito.doReturn(Flux.just(oldUnpublishedCollection.getName()))
+                .when(actionCollectionEntityRefactoringService)
+                .getExistingEntityNames(Mockito.anyString(), Mockito.any(), Mockito.anyString());
 
         final Mono<LayoutDTO> layoutDTOMono =
                 refactoringSolutionCE.refactorEntityName(refactorActionCollectionNameDTO, null);
