+++ app/server/appsmith-server/src/test/java/com/appsmith/server/widgets/refactors/WidgetRefactoringServiceCEImplTest.java
@@ -54,7 +54,8 @@ class WidgetRefactoringServiceCEImplTest {
 
     @BeforeEach
     public void setUp() {
-        widgetRefactoringServiceCE = new WidgetRefactoringServiceCEImpl(newPageService, astService, mapper);
+        widgetRefactoringServiceCE =
+                new WidgetRefactoringServiceCEImpl(newPageService, astService, mapper, pagePermission);
     }
 
     @Test
